fix problem with hashtags in name and comment
